sgpt/requirements.txt
torch>=1.10.0
transformers>=4.20.0
accelerate>=0.13.0
diffusers>=0.10.0
TTS>=0.8.0
moviepy>=1.0.3
numpy>=1.20.0
Pillow>=9.0.0
streamlit>=1.12.0
openai-whisper>=20230314
scipy>=1.7.0
soundfile>=0.10.3
requests>=2.25.0
sentencepiece>=0.1.96
